+++ Server/ServerCore/DBManager.h
@@ -10,13 +10,13 @@ public:
     void                            Clear();
 
     DBQueueRef                      GetDBQueue(int32 index);
-    DBQueueRef                      GetDBQueueFromId(int32 id);
+    DBQueueRef                      GetDBQueueFromId(int64 id);
     
     int32                           GetDBQueueCount() { return _dbQueueCount; }
 
 private:
     vector<DBQueueRef>              _dbQueueList;
     int32                           _dbQueueCount = 0;
-    std::hash<int32>                _hashGenerator;
+    std::hash<int64>                _hashGenerator;
 };
